GUI upgrade
- wind "angle" renamed to "effect"
- removed density and visconsity attributes
- removed patch size setup
- added TerrainPlanner button
- implemented extra patch pregeneration count setter

+++ Assets/Scripts/TerrainFunctions/GridManager.cs
@@ -20,8 +20,8 @@ public class GridManager {
     /// </summary>
     public Vertex GetPointOnGrid(Vertex point)
     {
-        int xVal = (int)(point.x / (stepX ));
-        int zVal = (int)(point.z / (stepZ ));
+        int xVal = point.x / stepX;
+        int zVal = point.z / stepZ;
 
         Vertex closestCenter = new Vertex(666, 666);
         for(int xi = -1; xi <= 1; xi++)
